feat: add jwt util
